third_party/WebKit/LayoutTests/fast/shapedetection/shapedetection-creation.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

// This test verifies that DetectedObject can not be created
test(function() {
  assert_throws(new TypeError(),
                function() {
                  var detectedObject = new DetectedObject();
                });
}, 'DetectedObject should not be instantiable.');

// This test verifies that Detector can not be created
test(function() {
  assert_throws(new TypeError(),
                function() {
                  var detector = new Detector();
                });
}, 'Detector should not be instantiable.');

// This test verifies that DetectedFace can be created
test(function() {
  var detectedFace = new DetectedFace();
  assert_true(detectedFace instanceof DetectedFace);
}, 'DetectedFace instance can be created.');

// This test verifies that FaceDetector can be created
test(function() {
  var faceDetector = new FaceDetector();
  assert_true(faceDetector instanceof FaceDetector);
}, 'FaceDetector instance can be created.');

</script>
